fix(server): gorm logging
- add support for LOG_LEVEL env var (Resolves #271)
